+++ apps/worker/sentry.ts
@@ -14,12 +14,12 @@ export const initWorkerSentry = () => {
         ],
 
         // Add Tracing by setting tracesSampleRate
-        // We recommend adjusting this value in production
-        tracesSampleRate: 0.005,
+        // Send 1% of transactions to Sentry
+        tracesSampleRate: env.SENTRY_TRACES_SAMPLE_RATE || 0.01,
 
         // Set sampling rate for profiling
         // This is relative to tracesSampleRate
-        profilesSampleRate: 0.001,
+        profilesSampleRate: env.SENTRY_PROFILES_SAMPLE_RATE || 0.01,
 
         beforeSend: (event) => {
             if (event.tags?.job && event.tags?.chain) {